requirements.txt
flask==2.2.3
werkzeug==2.2.3
pymongo==4.5.0
gunicorn==20.1.0
python-dotenv==0.19.2
pytest==7.3.1
dnspython==2.4.2
certifi==2024.2.2
# Additional dependencies for enhanced functionality
requests==2.31.0
Flask-Session==0.5.0
cryptography==41.0.3
Pillow==10.0.0
Jinja2==3.1.2
numpy==1.21.6
pandas==1.3.5
transformers==4.11.3
torch==1.9.1
scikit-learn==0.24.2
spacy==3.1.7
nltk==3.6.5
beautifulsoup4>=4.9.0
click>=8.0.0
Flask>=2.0.0
Jinja2>=3.0.0
MarkupSafe>=2.0.0
Werkzeug>=2.0.0
pyspellchecker>=0.6.2
regex>=2021.4.4
colorama>=0.4.4
python-dateutil>=2.8.2
pyyaml>=6.0
boto3>=1.20.0
dataclasses-json>=0.5.7 